Fix exporting a Sprite Sheet with Group name has different functionality between CLI and Scripting (fix #4456)
Before this fix, the lua command:
app.command.ExportSpriteSheet
could not process a layer within a group when the layer name
was expressed using the layer hierarchy path, for example:
layer = "Group1/Layer1"
