+++ src/app/util/layer_utils.cpp
@@ -75,4 +75,15 @@ bool layer_is_locked(Editor* editor)
   return false;
 }
 
+std::string get_layer_path(const Layer* layer)
+{
+  std::string path;
+  for (; layer != layer->sprite()->root(); layer=layer->parent()) {
+    if (!path.empty())
+      path.insert(0, "/");
+    path.insert(0, layer->name());
+  }
+  return path;
+}
+
 } // namespace app
