+++ src/app/util/layer_utils.h
@@ -1,5 +1,5 @@
 // Aseprite
-// Copyright (C) 2020  Igara Studio S.A.
+// Copyright (C) 2020-2024  Igara Studio S.A.
 //
 // This program is distributed under the terms of
 // the End-User License Agreement for Aseprite.
@@ -7,6 +7,8 @@
 #ifndef APP_LAYER_UTILS_H_INCLUDED
 #define APP_LAYER_UTILS_H_INCLUDED
 #pragma once
+
+#include <string>
 
 namespace doc {
   class Layer;
@@ -26,6 +28,8 @@ namespace app {
   // True if the active layer is locked (itself or its hierarchy),
   // also, it sends a tip to the user 'Layer ... is locked'
   bool layer_is_locked(Editor* editor);
+
+  std::string get_layer_path(const doc::Layer* layer);
 
 } // namespace app
